fix: default token logic
